ccimx6ul: determine part/vol to boot from on dual boot
A user can dynamically change to dual boot by setting variable
'dualboot' and running the install firmware script.
The variable 'mtdbootpart' may exist and be pointing to a name
that corresponds to the previous partition layout (like 'linux')
when it should be pointing to the new one ('linux_a' or 'linux_b').

Use the variable 'active_system' to set 'mtdbootpart' to the
appropriate part/vol when in dual boot.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DEL-7649