Revert "ubi: reserve blocks for bad block handling at the end of the NAND"
This commit may be useful when the single system partition
layout is used. In that case, all UBI volumes have a given size
except the last one, which uses the remaining space.

However, using the traditional Multi MTD layout, all UBI volumes
are created without specifying the size (so that they fill all
the space available in the MTD partition), so in this case,
the full amount of blocks is reserved per partition, which makes
our images (recovery, at least) not fit inside the volume.

By reverting this commit we return to just reserve a 4% of the
MTD partition size.

This reverts commit 003fcdce357bab220d0d8d62e11e10a658cd0270.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>