update_nand: add optional argument to force erase of partition
The 'update' command erases the MTD partition except when a ubifs
file is being updated to a UBI volume that already exists.
There are cases however, when we want to force the erase of the
MTD partition, for example when the partition table has been
modified, and UBI volumes may still exist in the NAND data, but
do not conform to the new partition table.

This commit adds a new optional parameter '-e' to the update command
that can only be added at the end and forces the erase of the
MTD partition before updating.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>