ccimx6ul: use variable 'rootfstype' to compose boot arguments
One of the actions that 'dboot' does is set the boot arguments.
It does so by running the script in 'bootargs_<media>_<os>'.
The current script for Linux needs to use different arguments when mounting
a UBIFS or a SQUASHFS root file system.
To do so, check the new variable 'rootfstype'. By default (if not set)
asume 'ubifs'.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>