sign: add variables for setting logs externally
Signature procedures for CC8X and CC8M platforms require the
use of logs generated when imx-boot is built.

Add the option to set the log paths via external variables. If not
provided, default logs will be expected to be in the folder where
the signing script is being called from.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>