ccimx6sbc: remove duplicity when setting mmc device
Historically we would set the mmc device using partition 0, and in case
it failed we would try again without any partition.

In the current U-Boot version, not selecting any partition defaults to
partition 0, so we are running the same command twice.

Remove the redundancy and so, simplify the partition scripts.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>