dboot: squashfs: define CONFIG_AUTH_SQUASHFS_ADDR for AHAB
To authenticate an artifact with AHAB, we first "load" it into a
position in RAM, and AHAB "decrypts" the file and moves it into
another position in RAM, defined during the signing process. This
happens even for only signed images.

If the image is too big and the loaded and decrypted images overlap,
AHAB can fail and reboot the device.

Define a new address 'CONFIG_AUTH_SQUASHFS_ADDR' and set it to 0x90000000
for AHAB builds by default so Squashfs files are loaded with
enough distance to the address used for decrypting (0x82100000) to not
overlap.

This allows for authentication of squashfs images of a maximum size of
223MB.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>