ccmp1: add HWID support as a stub based on cc8m
Create the needed header file with a preliminary HWID structure to
be able to build Digi HWID support for the CCMP1.
Create a folder 'ccmp1' and a common header for stuff that can be shared
by SOMs using variants of STM32MP1x SOC.
Create a Digi common Kconfig for Digi-specific configurations, such as the
SOM family, used in Digi code.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>