ccimx8m: move power definition out of common header file
The new CC8M Plus uses the new driver model for power initialization on
the SPL, so these configurations cannot be in the common file that will
be used also for the 8M Plus.

That said, the power configurations were already duplicated in the 8M
Mini and 8M Nano platform files, so this commit basically just removes
those definitions from the common file.

https://onedigi.atlassian.net/browse/DUB-1010

Signed-off-by: Javier Viguera <javier.viguera@digi.com>