ccimx8m: select minimum RAM size by default
We have observed that variants with less RAM that the default size can
fail to boot when the HWID has not been programmed. Select the minimum
RAM size by default so we ensure that any variant will work fine with
the default configuration.

That is 512MB for CC8M Nano and 1GB for CC8M Mini.

https://onedigi.atlassian.net/browse/MFGTEST-2091

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>