digi: separate authentication functionality from CONFIG_SIGN_IMAGE
Before this change, CONFIG_SIGN_IMAGE added three different things when
enabled:

    * Padding in the U-Boot image to fit a CSF
    * Boot artifact authentication support
    * For i.MX6 platforms, the generation of signed/encrypted U-Boot binares at
      the end of the build

When it comes to standalone signing, this posed a problem for i.MX6 platforms
because regular U-Boot images don't include boot artifact authentication
support. This forces users to enable CONFIG_SIGN_IMAGE to add said support, but
this also triggers the signing/encryption of the bootloader, making standalone
signing somewhat pointless.

Separate the signing process from the artifact authentication support to make
standalone signing more user-friendly. Note that unsigned i.MX6 bootloaders
with authentication support aren't meant to be run on a target, the point of
this change is simply to be able to separate the build process from the signing
process if needed.

For i.MX8 platforms, this change is transparent, as CONFIG_SIGN_IMAGE and
CONFIG_AUTH_ARTIFACTS have the same effect and the signing process is done at
the imx-boot level, not the u-boot level.

https://onedigi.atlassian.net/browse/DEL-7862

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>