net: dwc_eth_qos: enable clocks at probe() and don't stop them
The driver enables the clocks only at start(), which makes it impossible
to communicate with the PHY via 'mii' command.
Besides, the clocks are disabled after each network command, which again
makes it impossible to use any 'mii' command, and requires that each time
the PHY needs to autonegotiate, with the delay that it implies, and the
possibility of errors.

This patch:
- activates the clocks at probe()
- doesn't deactivate the clocks except on remove()

This allows for the use of 'mii' commands and avoids an auto-negotiation
with the PHY for every network command.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>