ccimx8mp: use NT6AN512T32AV DDR configuration
This is the memory used in the one prototype we have currently. The
configuration is forced to 16-bit even though the memory is a 32-bit
2GB size.

This is done to workaround a problem in the design with DDR_CS/1_B and
DDR_CKE0/1_B lines, preventing to use the LPDDR in 32-bit configuration.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>