mtd: stm32_fmc2_nand: use ECC strength reported by ONFI
This NAND driver starts assuming a default ECC strength
of 8 bits. If a different ECC strength and size is passed
from the DT, it then uses those. This may be ok for legacy
NAND chips, but it completely disregards the information
read from the ONFI table of the NANDs, for ONFI-compatible
NANDs. As such, it forces you to know in advance the ECC
parameters of the NAND and hardcode them on the DT.

This patch, instead, reconfigures the driver to use the
ECC layout as specified by the NAND chip via ONFI.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>