ccmp1: rework env_get_location() and move it to a common ccmp1.c file
This function returns the location of the environment basing on the
selected boot media. If the boot media (as selected by strapping) is not
one of the configured ones, it defaults to using no stored environment
(which will use the default one only in RAM).

On the CCMP1x, we want the SOM to use the default media (NAND) by default
no matter the selected boot media.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>