nand: set ECC strength to 1-bit for NANDs reporting 0 on ONFI ECC strength
Some NAND chips do not require ECC and report a 0 on ECC strength field on
the ONFI table. The NAND drivers, however, may not accept an ECC strength
of 0 and result in problems writing/reading data to/from the NAND.

Set the ECC strengh to 1-bit if the ONFI reports 0 ECC strength.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/CC6UL-1277