hwid: set fixed alignment to 32-bit words for hwid structure
Some architectures may align structures by default, however it is not required,
so due to our way of filling this structure in 32-bit words, we can run into
problems. This is the case with STM platforms where these structures are not
aligned and when we put them into words it fails.
This commit sets the hwid struct alignment for all Digi platforms.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>