ccmp1: remove 'singlemtdsys' variable and logic
On the CCMP1 SOMs, we only consider a single system partition
called UBI, with many UBI volumes.
The variable 'singlemtdsys' allows to change between this and
a system with multiple MTD partitions with one UBI volume each
(old approach on CC6UL).
This commit removes the variable and uses the logic as if it was
set.
For dboot and update, disregard the variable and look for UBI
volumes if an MTD partition can't be found.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>