cc6ul: fix conditional for setting of $mtdbootpart
In dual boot, the variable is set inconditionally to the active
system. In single boot, however, it was only set if it did not
exist before. The idea behind, is that this variable might
have been set to 'recovery', and in that case, it must not be
touched. However, if you transition from a dual boot system,
this variable may be set to 'linux_a' or 'linux_b', and then
we must set it back to 'linux'.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
(cherry picked from commit ad0f76ee2cc4b676e441338d46d286552236211d)
Signed-off-by: Javier Viguera <javier.viguera@digi.com>