mtdparts: reduce verbosity of find_dev_and_part() errors
This function is seldom used on U-Boot code (for nboot command
and dfu commands).
Though it is normally used with partition names, if a partition
name is not found it tries to parse a string in the form:
	<type><id>,<index>

and prints errors for incorrect device type (if not one of 'nand',
'nor'...) or if missing/wrong <id> or <index>.

We use this function in our 'update' command for updating UBI
volumes if an mtd partition with that name is not found, and this
produces the said error printfs, which are confusing.
Change the printfs to debug(), to reduce verbosity.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>