st: dfu: set 'dfu_alt_info' with every run
DFU mode might fail to initialize if the 'dfu_alt_info' variable is stored
in the environment with different information than the current:

  => dfu 0
  Card did not respond to voltage select! : -110
  Couldn't init MMC device.
  DFU entities configuration failed!
  (partition table does not match dfu_alt_info?)

This could happen if 'dfu' is run with a predefined 'dfu_alt_info' that
contains information about the microSD card partitions but microSD card
is no longer inserted.

To avoid it, reset 'dfu_alt_info' with every 'dfu' run.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>