sentinel: fuse: do not error when value to program is zero
Current code exits with -EINVAL when the value to program in the fuses
is zero. This leads to an error in our hwid command when one of the words
is zero.

For example:

=> hwid prog 00007000 00000000 5ae00005
 ...
Warning: Programming fuses is an irreversible operation!
         This may brick your system.
         Use this command only if you are sure of what you are doing!

Really perform this fuse programming? <y/N> y
Programming HWID... ERROR
exit not allowed from main input shell.

Instead, do nothing when the value is zero, but give a success return
value to the consumers.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>