update: let use 'linux'|'rootfs' keynames when dualboot enabled
Standard names for system partitions in DEY are 'linux' and 'rootfs'.
With the enabling by default of dualboot, the system partitions
are named instead 'linux_a'/'linux_b' and 'rootfs_a'/'rootfs_b'.

This commit checks if dualboot is enabled and if the partition name
to update is 'linux', resolves to update the active partition (whether
linux_a or linux_b) and the same with 'rootfs'.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DUB-1041