hwspinlock: support hwlock-cells = 2 in default binding
Since Linux commits "hwspinlock: allow sharing of hwspinlocks",
the simple translation function change to support 2 parameters.

If a device registers to the framework with #hwlock-cells = 2, then
the second parameter of the 'hwlocks' DeviceTree property defines
whether an hwlock is requested for an exclusive or a shared usage.

Even if this parameter is not supported in U-Boot hwspinlock uclass,
this patch adds support of device tree with this new parameter.

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: I99ade8215b846d9a67085f988800071754f6cd8f
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/285578
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>