defconfig: reduce DDR cache size to 128MB to solve optee permission issue
This commit reduces the u-boot cachable DDR size to 128MB to solve an optee permission
issue on the CCMP13. Normally the first 256MB of the SRAM is cacheable and Uboot pre-reload
is executed at SDRAM bottom which is defined by optee as a secure area.

Solution is to reduce the cacheable area size to 128MB.

Signed-off-by: Mike Engel <Mike.Engel@digi.com>