ccimx8: make loadscript variable more robust
In dual-boot mode, if `active_system` and `mmcpart` are not in sync,
the boot script may be read from the incorrect boot partition. So make
the `loadscript` default value more robust, by using `active_system` as
the single point of truth to decide from which partition the boot script
should be loaded from.

https://onedigi.atlassian.net/browse/DUB-1048

Signed-off-by: Javier Viguera <javier.viguera@digi.com>