ccimx93: environment: simplify bootcmd_mfg command
When the device boots from USB, bootcmd_mfg is executed. The default
value is more complicated than what we need, as it checks images in
memory (with 'iminfo' command) and tries to boot them automatically.

So simplify the command to go to fastboot mode, just the same as the
rest of our platforms.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>