ccimx93: correct clock and strobe pad setting
Correct the configuration of clk and strobe pads for the usdhc interfaces,
which were configured as pull up/down and must be configured (both) as
pull-downs, otherwise it may cause problems with some eMMCs when enabling
the HS400ES mode, with CMD8.

This is a backport of mainline kernel commit:
62f0147fd4d86620853bee027800f988d3013656

On the CC93 this was causing issues with Foresee eMMCs.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>