ARM: dts: stm32: add sdmmc2 & 3 nodes on stm32mp257f-ev-revB
The sdmmc2 node is used for the eMMC embedded on the board.
The sdmmc3 node is used to connect an SD-card with an extension board
connected to RPi GPIO expansion connector. As the card is not present
by default, the node status is disabled.

Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Change-Id: I11b73918142db51d4f4f83d84357afd0ef2c27ba
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/308426
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Patrice CHOTARD <patrice.chotard@foss.st.com>
Domain-Review: Christophe KERELLO <christophe.kerello@st.com>