ccimx8x: restore old defconfigs to ensure backwards compatibility
Older DEY versions are still using this version of U-Boot, and if we rework the
way configurations are organized by RAM size, we're going to break
compatibility with said DEY versions. We can make the new system co-exist with
the old one by restoring the previous defconfigs as they were and including
conditional compilation.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>