ccimx8x: change variant info sdram type to phys_size_t
At the moment, we only use this field in a corner case when determining the
target's RAM size. If the HWID is programmed but the RAM size field is not
coded or has an unknown value, we parse the variant ID and obtain the
corresponding RAM size (sdram field) from the variant table.

However, this field was using the signed "int" type, and when using a size of
2 GiB (0x80000000), it's being interpreted as a signed number, so technically,
it's -2GiB. When casting this value to a 64-bit type (phys_size_t, which is
unsigned long long), it does so as a signed number, and the leftmost bits are
populated with 1s (0xFFFFFFFF80000000). This is then stored in an unsigned
variable and interpreted as such. As you can imagine, this leads to undefined
behavior early on in the boot process, causing U-Boot to crash and reset
indefinitely.

Change the sdram field type to phys_size_t to avoid this, and also to prepare
for 4GiB variant support in the near future.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>