ahab: spl: copy DEK blob from 1st container to 2nd container to decrypt U-Boot
Until now, when using encrypted imx-boot images on platforms using AHAB
(ccimx8x), we would store the DEK blob in one of the containers in the image.
Now that we are using images with SPL that have an additional container, and
said container also requires a DEK blob stored within it, encrypted images
don't boot.

We could modify the logic to store the DEK blob so that it gets stored in both
containers, but this breaks backwards compatibility: when flashing a new
encrypted imx-boot image from an old one, the DEK blob would only get stored in
the first container. Instead, we keep this logic intact and copy the DEK blob
from the 1st container to the 2nd one on the fly while the SPL is running,
right before the 2nd container gets authenticated by the SECO.

This requires the following changes:

    * When parsing the container headers to determine the offset at which
      U-Boot is stored in the storage media, read 80 bytes of additional data.
      This is because the DEK blob is stored at the very end of the container
      header and it overruns its default boundaries.
    * When the first container is parsed and it contains a legitimate DEK blob,
      store it in a buffer.
    * Before authenticating the second container, check that it contains a blob
      offset (this indicates that a blob is suppossed to be in the container),
      but no actual blob is stored there. If so, copy it from the
      aforementioned buffer.

This allows encrypted imx-boot SPL images to boot on closed targets.

https://onedigi.atlassian.net/browse/DEL-8620

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>