ccimx8x: use a maximum of 2 GiB for update operations
The logic we have to calculate the verify address in update operations assumed
that the complete RAM was available in a single, continuous address space. This
isn't true for ccimx8x variants with 4 GiB of RAM (1st 2 GiB are available
between 0x80000000 and 0xFFFFFFFF, and the remaining 2 GiB are between
0x880000000 and 0x8FFFFFFFF). Because of this, the verify address was being set
in an inaccessible address between both banks, triggering an abort and a reset
during verification.

If we work around this by moving the verify address to the second SDRAM bank,
we still trigger an abort because NXP's MMC driver uses 32-bit DMA addresses,
and the second SDRAM bank requires 64-bit addresses.

Use only the first SDRAM bank for update operations. This limits 4 GiB variants
to use only 2 GiB during updates, but we have the on-the-fly mechanism
available for larger updates.

https://onedigi.atlassian.net/browse/DEL-8604

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>