bootcount: configure 'bootcount' to use MCA NVRAM register in the ConnectCore products
Modify the integration of the 'bootcount' feature in the ccimx6ul, ccimx8x and ccimx8m family
devices to use a MCA NVRAM registers. From now, the boot count value will be stored in the MCA
instead of the U-Boot environment, allowing the bootcount feature to be always enabled and not
only when the 'upgrade_available' variable is set to 'yes'.

The bootcount uses 2 bytes: 1 byte for the magic (0xBC) and one byte for the real bootcount. This
means that only 2 of the 8 bytes of the MCA NVRAM are being used with this functionality.

For ccimx6ul products, the 'CONFIG_SYS_I2C' define had to be removed in order to use the device model
device tree mechanism, as the 'CONFIG_DM_I2C' and 'CONFIG_SYS_I2C' symbols cannot co-exist.

While on it, enable the 'bootcount' command to be able to read and reset the 'bootcount' value from
U-Boot.

https://onedigi.atlassian.net/browse/DEL-8506

Signed-off-by: David Escalona <david.escalona@digi.com>