config: remove unnecessary rootfstype variable from config
We can get the image type reading the header of the rootfsvol
and checking the magic number of squashfs.

We can't use loadaddr to the commands because in this step
the bootscript/kernel/dtb are already loaded. We are using
update_addr that is free to this kind of operations.

Also the rootfstype parameter is not needed in read only filesystems
the kernel is able to detect it and mount it as ro.

The kernel is able to find the ubi block using the index, and the name
of the volume, use it instead of hard-coded value.

Signed-off-by: Francisco Gil <francisco.gilmartinez@digi.com>