bootcount: configure 'bootcount' to use 'User RAM' from external RTC
Modify the integration of the 'bootcount' feature in the CCMP1 family devices to use the external RTC
device. From now, the boot count value will be stored in the 'User RAM' registers of the development
board RTC device instead of the U-Boot environment. This allows the bootcount feature to be always enabled
and not only when the 'upgrade_available' env variable is set to 'yes'.

The external RTC device used in the development board (Micro Crystal RV-3028) provides 2 general purpose
'User RAM' registers to store data and backup them with battery. These registers are also protected against
soft resets, so they are the perfect place to store values between reboots. The address of these registers
in the RTC chip are 0x1F and 0x20.

While on it, enable the 'bootcount' command to be able to read and reset the 'bootcount' value from U-Boot.

https://onedigi.atlassian.net/browse/DEL-8506

Signed-off-by: David Escalona <david.escalona@digi.com>