ccimx6ul: adapt 'altbootcmd' for single boot systems
Until now, the 'altbootcmd' assumed the system was always a dualboot system. This is
because the 'bootcount' mechanism was only active for this kind of systems. With the recent
changes, the 'bootcount' mechanism is now always active, it does not matter if the system is
single or dual boot. This scenario was causing the boot process to break in single boot
systems if the 'bootcount' limit was reach and the 'altbootcmd' was executed, because single
boot systems do not have 'linux-a' partition and the script was looking for it. This commit
checks first if the system is dual boot to look for the 'linux-a' partion or single boot to
look for the correct 'linux' partition instead.

https://onedigi.atlassian.net/browse/DEL-8506

Signed-off-by: David Escalona <david.escalona@digi.com>