ccmp1: rework partition layout and disable UBI fastmap
Up to now, one MTD partition called "UBI" contained all system UBI volumes,
including the U-Boot environment itself.
U-Boot needs to attach this unique UBI device, which is very time consuming
due to the scanning of the full partition, which is large.
For this reason, we were enabling the Fastmap option, that after a first
scan uses a fast lookup table. This is an experimental mechanism that has
apparently discontinued development.

Disabling Fastmap has the disadvantage of the amount of time U-Boot needs
to attach the UBI partition. To minimize this time, this commit reworks
the NAND partition layout to have two partitions:
- One small "UBI" partition, to hold the U-Boot environment (and redundant
  copy) plus the Linux volumes (that hold kernel, DTs, bootscripts...)
- One large "UBI_2" partition, to hold the rootfs and data volumes.

U-Boot only needs to attach the small "UBI" partition to access the
environment and the Linux boot artifacts, which takes an affordable amount
of time.

As part of this commit, implement a Digi custom board_mtdparts_default()
that creates the new partitions basing on the available NAND size.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>