ccmp1: move MTD_UBI_MAXRSVDPEB_PCNT to Digi Kconfig
The mechanism to limit the amount of blocks reserved to bad block handling
was introduced by Digi in commit 16c52cd409a1287cd4c1deea5c02ea1939d86e87
but the configuration was hardcoded in board include files.
This commit moves it to Digi Konfig and sets values on board defconfigs.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>