update_nand: look for UBI volumes in one or more system partitions
With the disabling of fastmap mechanism we need to create two MTD system
partitions (UBI, UBI_2) so that the attach of the partition with the boot
artifacts doesn't take very long. The rootfs volumes are then moved to
UBI_2 but we need to add support to the 'update' command to look for
volumes to update.
This commit adds this support by iterating over N system partitions that
must be named with suffix "_2", "_3" and so on.

https://onedigi.atlassian.net/browse/DEL-8652

Signed-off-by: Hector Palacios <hector.palacios@digi.com>