ccimx93-dvk: configure RTC's backup switching mode
This is in the context of using the RTC to save the boot counter.

In the ccimx93-dvk, the RTC loses power on every reset (software
or hardware), and thus, the boot counter is lost. The RTC relies on a
battery backup (in our case a coin cell), but that requires changing the
RTC configuration in the EEPROM. By default the backup switching mode
(BSM) is set to zero (which means switchover is disabled), this commit
sets the BSM to level switching mode (LSM).

Signed-off-by: Javier Viguera <javier.viguera@digi.com>