albootcmd: integrate 'altboot' script functionality in 'altbootcmd' for cc6 platforms
When the system is unable to boot during 3 consecutive attempts, U-Boot executes the ‘altbootcmd’ to
initialize the alternative boot process. This command relies in the contents of the boot partition as
it executes the ‘altboot’ U-Boot script located in that partition. The script swaps the active partitions
bank (in the case of dualboot systems), resets the boot count and reboots the system. The full process
works fine if access to the boot partition is guaranteed, but it will definitely fail if the boot
partition is not available because of memory corruption or there are no contents in that partition.

This script removes the dependency with the boot partition to execute the ‘altboot’ script by moving that
functionality directly to the ‘altbootcmd’ in U-Boot. This way, if the boot partition is corrupted or does
not have the correct contents, the alternative boot process can continue without issues.

https://onedigi.atlassian.net/browse/DEL-8674

Signed-off-by: David Escalona <david.escalona@digi.com>