dboot/update: fix use of default dev:part when using mmc
When using external mmc to read files from, and not providing a specific
dev:part indexes, the values of 'mmcdev' and 'mmcpart' env variables
should be taken instead.
The function that resolves this is get_default_devpartno() but it bases
on the value of 'fwinfo->src' which was not set until a later point.

Set 'fwinfo->src' right after the loop where it is determined to fix
this behavior.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
(cherry picked from commit 78f675c852904d18e2dcb9378bc2d706ae6275b9)
Signed-off-by: Javier Viguera <javier.viguera@digi.com>