cmd: trustfence: encapsulate functions to fix build failures
Not all the functions on this file are supported by all platforms. Some
functionality is for old HAB4-based devices, some functions are weak
implementations, etc. The result is that build failures happen when you
start implementing TF for a new platform (for example the AHAB-based
ccimx93)

This commit tries to encapsulate some functions and provide stubs so the
code is buildable when not all the CONFIG_TRUSTFENCE macros are defined.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>