boot.kconfig: remove SIGN_IMAGE dependence on AUTH_ARTIFACTS
AUTH_ARTIFACTS is a u-boot functionality to authenticate "external"
artifacts (kernel, dtbs, etc.) and SIGN_IMAGE is about signing the
u-boot image itself. So you may want to enable SIGN_IMAGE without
AUTH_ARTIFACTS. For that remove the dependence.

On the other hand, authenticating artifacts with a u-boot that is not signed
does not make sense, so select SIGN_IMAGE when AUTH_ARTIFACTS is selected.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>