build.sh: shorten toolchain installation path
After unpacking, yocto toolchains do a binaries relocation
post-processing. This is good because it allows installation of the
toolchain in a different path to where it was built/generated.

There is a limitation, though. As the relocation patches the binaries by
changing the interpreter (loader), the new interpreter path (which depends
on the installation directory) cannot be longer than the original one in
the binary. If it is longer than the original, the relocation fails with:

ERROR: could not relocate ... interp size = X and Y is needed.

See:
https://lore.kernel.org/all/CAFQ+DijY3q_Fc56uwCKuH_hPiH1LMk3_B=SR5rqWpA70OHxZ7g@mail.gmail.com/T/#u

This commit saves a few characters in the (so nested) toolchain
installation path to minimize such failures.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>