ccmp1x: disable STM32PROG over UART
This configuration allows programming partitions over UART
using STM32_Programmer_CLI, but has the disadvantage that it
also disables the serial console, which prevents to just
recovery boot from the same UART used as U-Boot console.

On the CCMP15-DVK, the UART used as console (UART4) uses the
IOMUX of the pads expected by the ROM loader. With this change
the console can be used as a recovery boot method.

On the CCMP13-DVK, the UART used as console (UART5) uses different
IOMUX and it cannot be used for recovery boot, but disable the
setting anyway, to be in line with the CCMP15. Digi U-Boot already
provides different methods to update the firmware.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>