autoboot: skip bootcount increment if bootlimit=0 or not defined
Setting 'bootlimit' to 0 (or deleting the variable) has the effect of
disabling the actions when bootcount reaches the limit, but doesn't disable
the increment of the bootcount.
This can be confusing (bootcount increments without limit) apart of using
resources (writing to NVRAM) without reason.

This commit skips the bootcount increment if the limit has been disabled,
disabling completely the mechanism.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DUB-1058