digi: set bootlimit to 0 to disable the boot count mechanism by default
Enabling the bootcount mechanism shall be a conscious action that's taken
when there are guarantees about the actions about the deployed system and
the actions to be taken (having a dual system deployed, a watchdog running,
an altbootcmd...).

Enabling this mechanism by default may lead to unexpected actions (swap of
active system, reset) during manufacturing processes or development
evaluation.

Re-enabling the mechanism is as easy as setting a positive bootlimit.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DUB-1058
(cherry picked from commit 4d290dfd120fda19e0526fcbfde8310f31838734)