digi: restore code to get filename basing on 'dboot_kernel_var'
The variable 'dboot_kernel_var' contains the variable that
contains the default filename for the kernel.
If it contains 'zimage' then the variable 'zimage' contains
the default name for a zImage kernel.
If it contains 'uimage' then the variable 'uimage' contains
the default name for a uImage kernel.

For FIT images, this variable should be set to 'fitimage'
since this is the variable that contains the default name
for FIT images.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>