dboot: skip re-loading fitimage if temp variable set
The FIT image may contain a script that itself runs the 'dboot' command.
In that case, dboot would load again the fitImage into memory, when it
is already there (previously loaded to run the boot script itself).
To avoid this double loading, check a temporary env variable
'temp-fitimg-loaded', that the script may set.
If set, skip the re-loading of the fitImage, print a message, and clear
the variable.

While on it, change the dboot message "## Loading kernel" to print instead
"## Loading fitImage" when using FIT. Internally, the fit will print a
message with every artifact that it reads.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>