cmd/booti: simplify Trustfence artifacts authentication code
The digi_auth_image generalization for HAB/AHAB complicates the code
here. For AHAB devices, the function needs to run BEFORE "booti_setup"
as it does an image relocation that booti_setup needs to parse the ARM64
kernel header. On the other hand, for HAB devices, the booti_setup needs
to be executed before "authenticate_image" to get the image size from
the kernel header and use it for image authentication.

So use the "digi_auth_image" wrapper only for the AHAB case and pass a
dummy 0 in the image size as for AHAB is not used. This allows cleaning
up the code and minimizes changes compared to the NXP-released code,
which eases maintenance and future migrations.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>